move tests to test to avoid bad autocomplete
