+++ test/auth.js
