+++ test/share.js
